refactor: 주문하기 기능에 대한 핸들러 함수 정의를 커스텀 훅에서 컴포넌트로 이동
